requirements.txt
python-dotenv
requests
python-socketio[client]

# Images
opencv-python
gpiozero